requirements.txt
baal
fontTools==4.4.0
h5py==2.10.0
matplotlib==3.1.3
numpy==1.18.1
pandas
pycairo==1.19.1
pyicu-binary==2.3.1
Pillow
requests
scipy
tensorboardX
torch
torchvision
--extra-index-url https://test.pypi.org/simple/
synbols==0.0.1.dev3
haven-ai
git+https://github.com/NVIDIA/apex
